+++ contracts/stake/src/lib.rs
@@ -1,6 +1,7 @@
 #![no_std]
 mod contract;
 mod distribution;
+mod error;
 mod msg;
 mod storage;